fold line data type

+++ src/App/objects/geometries/FoldLineGoemetry.ts
@@ -1,7 +1,7 @@
 import * as THREE from "three"
 
 const foldLineGeometry = new THREE.BufferGeometry()
-const foldAttribute = new THREE.BufferAttribute(new Float32Array(3 * 4), 3)
+const foldAttribute = new THREE.BufferAttribute(new Uint8Array(3 * 4), 3)
 
 export function newFoldLineGeometry() {
     const fl = foldLineGeometry.clone()
